Struct liquidity_api::ElectionResolvers

pub struct ElectionResolvers { /* fields omitted */ }

Methods

impl ElectionResolvers

pub fn new(cache_capacity: usize, cache_ttl: Duration) -> ElectionResolvers

pub async fn create_election<T, C>(
&'_ self,
input: ElectionInput,
context: &'_ C
) -> Result<Election, Box<dyn Error + 'static>> where
C: Context<T>,
T: DbConnection,

Create a new election

Arguments

input - The election user input object context - The request context, passed automatically

Permissions Required

create:election

Returns

The new Election or an error if the creation failed

Example

mutation {
    createElection(input: {name: "test"}) {
       id
       name
       description
    }
}

pub async fn edit_election<T, C>(
&'_ self,
id: Uuid,
input: ElectionInput,
context: &'_ C
) -> Result<Election, Box<dyn Error + 'static>> where
C: Context<T>,
T: DbConnection,

Edit an election

Arguments

id - The id of the election to be edited input - The fields to edit context - The request context, passed automatically

Permissions Required

update:election

Returns

The edited election or an error if the update failed

Example

mutation {
    editElection(id: "54c8ed41-b4f4-4f1c-8903-f9bbbe2d992d", input: {description: "new description"}) {
       id
       name
       description
    }
}

pub async fn election<T, C>(
&'_ self,
id: Uuid,
context: &'_ C
) -> Result<Option<Election>, Box<dyn Error + 'static>> where
C: Context<T>,
T: DbConnection,

Fetch an election by id

Arguments

id - The id to look up the election by context - The request context, passed automatically

Permissions Required

view:election

Returns

The election if it exists, None if it doesn't, Error if an issue has occurred

Example

query {
    election(id: "some_uuid") {
        id
        name
        choices
    }
}